sensor: Fix the BPF string
Before this change, the default BPF filter for sensor configured with
the server output was ignoring the server's port entirely. Which is a
bug if the sensor host has some server on the same port (and it should
be captured).

This change fixes that by specifying the server IP as a destination.

Signed-off-by: Michal Rostecki <vadorovsky@gmail.com>