streamer(common): Don't require an fd when plugin is enabled
Before this change, PacketStreamer was always expecting some file
descriptor to be available. That was forcing users wanting to output
only to plugin(s) to configure a /dev/null file anyway. Then
PacketStreamer was performing a wasteful job of writing to /dev/null.

This change fixes that by allowing to configure only plugins and not
writing to fd when it's not present.

Signed-off-by: Michal Rostecki <vadorovsky@gmail.com>