Add timeout to send captured packets early
Prior to that change, packets were only sent when the gather capacity
was reached. On low traffic, this means data is sent very rarely (and
might end up being lost).

The proposed change adds a configurable timeout mechanism.
If the timeout is reached before the capacity is full, captured packets
are sent over to the receiver.